refactor : application 에서 진행되던 게임을 Game 클래스 별도 생성으로 진행
Application 에서는 구동만 처리하고
실제 게임 구동 프로세스는 Game 클래스에서 진행하도록 처리
